refactor(styled.js): Move all styling to styled.js. Use Styled.<Component> notation.
